Dear Log, Motivation for the SVN foundation we'd like to (by the way, Anna thinks Jeff might be gay...but Audrey always thought I was, especially because of the whole rowing thing) anyway, where was I.
Tried to merge branch/tesFolderRev98/Code - Source into trunk/Code - Source, and failed.  When we tried to integrate branch back into trunk, it choked on some gaaaaayyyy ansestrally error.  

Sooo, we decided it was failing because 
a) we fucked up when we branched..
b) the file structures aren't the same between rev 98 in branch, and rev 105 in trunk,,, 

so we hand jammed the changes we wanted. and now we're going to try to merge again...weeeeeehheeeheheeee