adding Reference - xBee Profiles directory with the factory default profile saved from wire antenna xBee s/n 4067949E, and a nominal DogFighter xBee profile (factory default + 38400 + 555 VID).
i was tearing my hair out because the radios worked just fine with PuTTY magic text test, but failed once in the DF prototypes. turns our there's a mechanical engaugement problem with RPSMA xBee s/n 004068B516 in DF1 beta.  lifting the front (antenna) side just *slightly* out of the red sparkfun breakout board made everything work.  before that, it was not receiving data properly.  the problem is likely with the *combination* of that radio and that breakout board.  swapping radios between DF2 and DF1 solved the problem.  (that is s/n 004068B516 on DF2, and s/n 4067949E on DF1.)  

that shit's CRA.