another incremental check in.  added a Stopwatch class to trunk/Sandbox/SDCardHellowWorld.  the results of this run of logging are in trunk/Sandbox/DeckerKicksNetduinoAss.txt
there's three columns: text (filler data), row number, and ElapsedMilliseconds.

each row is between 4096 and 4736 bytes of data (assuming 128-byte ascii encoding...its possible .net is using 256!) depending on the row number. 

the upshot is this: using this implementation it takes ~30ms to write a row to the SD card.  that's a long ass time!

next steps/suggestions:
- pretty sure this means we'll have to log in binary (hex) and not ascii (sorry no pulling a .csv off the SD card...) so we should log a more appropriate amount of data: whatever two AttNavs are (me and enemy)
- play with opening/closing/flushing both the streamWriter to see if that has effects.  somehow the streamWriter has to load the data to memory before logging--how good is it at cleaning up its memory space?
- figure out why the red squigglies won't go away in Lowell's branch.
- tits
- profit