check in for SDCardHelloWorld.
changed where the stopwatch started in the code.  previously the stopwatch started before mounting the SD card and starting a filestream and a steamwriter; thus the average for-loop times were burdened with this overhead. this rev moves the stopwatch start and stop immediately before and immediately after the for-loop.

the per 1000-line times are noted in the comments of the code (e.g. "2500ms/1k-line") for various data types.  looks as though there's some advantage to writing binary (hex) data to the card, the the convince of writing ascii data probably outweighs the performance gain.  DeckerKicksNetduinoAss.txt has been updated. the single column shows elapsed milliseconds using this rev of code.

the so-what: we can expect SD card logging to take between 1 to 3ms per entry.