adding more ground test data.
minor changes to the logging scheme.  added a (lazy ass implementation) of a DLC to the Logger.

need to change the debug serial output to show the user TSA and yawWithMagCal_mrad so one can trying to drive those two numbers to be the same.  (presently outputting plain 'ol yaw.  if you do a magCal, you're not sure where the DF actually thinks it's pointing.)