merged branch/clockCycleSdCardLogging/DogFighter back into  trunk/Code - Source/DogFighter
I've used Notepadd++'s compare function to verify that all of the altered files are the same now in trunk as they were in the branch/clockCycleSdCardLogging folder, so the build should run just as it did before I got my mitts on it.


Also, please be sure that your working copy is the trunk/Code - Source/DogFighter code for future updates and commits.